requirements.txt
PyPDF2>=3.0.0
Pillow>=9.0.0

